% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynlib.R
\docType{data}
\name{coalMining}
\alias{coalMining}
\title{Coal Mining Disasters data}
\format{A \code{data.frame} with 112 entries, corresponding to the
years from 1850 to 1962. Each entry has:
\describe{
  \item{accidents}{ number of severe accidents that have occurred each year.}
  \item{years}{ year during which accidents occurred.}
}}
\usage{
data(coalMining)
}
\description{
A dataset with 112 entries containing annual numbers of accidents due to disasters in
British coal mines for years from 1850 to 1962, considered in Carlin et al. (1992) and
also in Eilers and Marx (1996).
}
\references{
Carlin, B.P., Gelfand, A.E. and Smith, A.F.M. (1992).
Hierarchical Bayesian analysis of changepoint problems.
\emph{Applied Statistics}, \strong{41}(2), 389--405.

Eilers, P.H.C. and Marx, B.D. (1996). Flexible Smoothing with B-splines
and Penalties. \emph{Statistical Science}, \strong{11}(2), 89--121.
}

