% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods_GeDSboost-GeDSgam.R
\name{coef.GeDSboost}
\alias{coef.GeDSboost}
\alias{coefficients.GeDSboost}
\alias{coef.GeDSgam}
\alias{coefficients.GeDSgam}
\title{Coef method for GeDSboost and GeDSgam objects}
\usage{
\method{coef}{GeDSboost}(object, n = 3L, ...)

\method{coefficients}{GeDSboost}(object, n = 3L, ...)

\method{coef}{GeDSgam}(object, n = 3L, ...)

\method{coefficients}{GeDSgam}(object, n = 3L, ...)
}
\arguments{
\item{object}{the  \code{\link{GeDSboost-class}} or \code{\link{GeDSgam-Class}} object from which
the coefficients should be extracted.}

\item{n}{integer value (2, 3 or 4) specifying the order (\eqn{=} degree \eqn{+ 1}) of the GeDS fit
whose coefficients should be extracted. If \code{n = 2L} piecewise polynomial coefficients of univariate
base-learners are provided. In the case of bivariate base learners,
the B-spline coefficients for each iteration on which such base-learner was selected are provided.
If \code{n = 3L} or \code{n = 4L} B-spline coefficients are provided.
By default equal to \code{3L}.
Non-integer values will be passed to the function \code{\link{as.integer}}.}

\item{...}{potentially further arguments (required by the definition of the generic function).
They will be ignored, but with a warning.}
}
\value{
A named vector containing the required coefficients of the fitted multivariate predictor model.
The coefficient of linear base-learners is named as the base learner itself. The coefficients of GeDS
base-learners are named as the base learners themselves, followed by the index of the corresponding B-spline.
}
\description{
Methods for the functions \code{\link[stats]{coef}} and \code{\link[stats]{coefficients}}
 that allow to extract the estimated coefficients of \code{\link{GeDSboost-Class}} 
 or \code{\link{GeDSgam-Class}} object.
}
\details{
See \code{\link{coef.GeDS}} for details.
}
\seealso{
\code{\link[stats]{coef}} for the standard definition; \code{\link{NGeDSboost}}
and \code{\link{NGeDSgam}} for examples.
}
