% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods_GeDSboost-GeDSgam.R
\name{deviance.GeDSboost}
\alias{deviance.GeDSboost}
\alias{deviance.GeDSgam}
\title{Deviance method for GeDSboost and GeDSgam objects}
\usage{
\method{deviance}{GeDSboost}(object, n = 3L, ...)

\method{deviance}{GeDSgam}(object, n = 3L, ...)
}
\arguments{
\item{object}{the \code{\link{GeDSboost-class}} or \code{\link{GeDSgam-class}} object from which
the deviance should be extracted.}

\item{n}{integer value (2, 3 or 4) specifying the order (\eqn{=} degree \eqn{+ 1}) of the
GeDSboost/GeDSgam fit whose deviance should be extracted.
By default equal to \code{3L}.
Non-integer values will be passed to the function \code{\link{as.integer}}.}

\item{...}{potentially further arguments (required by the definition of the generic function).
They will be ignored, but with a warning.}
}
\value{
A numeric value corresponding to the  deviance of the selected GeDSboost/GeDSgam fit.
}
\description{
Method for the function \code{\link[stats]{deviance}} that allows the user to extract the value
of the deviance corresponding to a selected GeDSboost or GeDSgam fit from a
\code{\link{GeDSboost-Class}} or \code{\link{GeDSgam-Class}} object.
}
\details{
This is a method for the function \code{\link[stats]{deviance}}.
As \code{\link{GeDSboost-class}} and \code{\link{GeDSgam-class}} objects contain three different
fits (linear, quadratic and cubic), it is possible
to specify the order of the GeDS fit for which the deviance is required via the input argument \code{n}.
}
\seealso{
\code{\link[stats]{deviance}} for the standard definition;
}
