% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynlib.R
\docType{data}
\name{EWmortality}
\alias{EWmortality}
\title{Death counts in England and Wales}
\format{A \code{data.frame} with 109 entries and 3 variables: \code{Age},
  \code{Deaths} and \code{Exposure}.

  {\code{Exposure} is a mid-year estimate of the population exposed to risk.}}
\usage{
data(EWmortality)
}
\description{
The dataset consists of information about the mortality
of the English and Welsh male population aggregated over the years 2000, 2001 and 2002.
}

