% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R, R/S3methods_GeDSboost-GeDSgam.R
\name{family.GeDS}
\alias{family.GeDS}
\alias{family.GeDSboost}
\alias{family.GeDSgam}
\title{Extract Family from a GeDS Object}
\usage{
\method{family}{GeDS}(object, ...)

\method{family}{GeDSboost}(object, ...)

\method{family}{GeDSgam}(object, ...)
}
\arguments{
\item{object}{A \code{\link{GeDS-class}} object.}

\item{...}{further arguments (ignored).}
}
\value{
An object of class \code{\link[stats]{family}} describing the 
distribution and link function used in the GeDS fit.
}
\description{
Method for \code{\link[stats]{family}} that returns the error distribution 
family used in the fitted GeDS model.
}
\seealso{
\code{\link[stats]{family}}, \code{\link{GeDS-class}}
}
