% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeDSClass.R
\docType{class}
\name{GeDSboost-class}
\alias{GeDSboost-class}
\alias{GeDSboost-Class}
\title{GeDSboost Class}
\description{
A class representing a fitted GeDSboost object. This object encapsulates details about the
GeDS gradient boosting model, including its formula, arguments, models, the final model,
predictions, and internal knots. Objects of this class are typically returned by the \code{\link{NGeDSboost}} function.
}
\section{Slots}{

\describe{
\item{\code{formula}}{A formula object representing the relationship between the variables used in the model.}

\item{\code{args}}{A list containing the arguments passed to the \code{\link{NGeDSboost}} function. 
It includes details about base learners, family, initial learner, knots for initialization, normalization options, outcome, 
predictors, shrinkage parameter, mean and standard deviation of predictors (X_mean, X_sd) and outcome (Y_mean, Y_sd).}

\item{\code{models}}{A list containing the models generated during the boosting process.}

\item{\code{final_model}}{A list detailing the final GeDS model after all boosting iterations. This includes 
base learners, the best base learner (best_bl), cubic fit, model name, quadratic fit, and predicted values (Y_hat).}

\item{\code{predictions}}{A list containing predictions obtained from the model for different fit types 
(linear, quadratic, and cubic).}

\item{\code{internal_knots}}{A list detailing the internal knots used for the different fit types 
(linear, quadratic, and cubic) during the boosting process.}
}}

\references{
Dimitrova, D.S., Kaishev, V.K., Lattuada A. and Verrall, R.J. (2017).
Geometrically designed, variable knot splines in Generalized (Non-)Linear Models.
Available at \href{http://openaccess.city.ac.uk/18460/}{openaccess.city.ac.uk}
}
\seealso{
\code{\link{NGeDSboost}} which generates objects of this class.
}
