% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods_GeDSboost-GeDSgam.R
\name{N.boost.iter}
\alias{N.boost.iter}
\alias{N.boost.iter.GeDSboost}
\title{Extract Number of Boosting Iterations from a GeDSboost Object}
\usage{
\method{N.boost.iter}{GeDSboost}(object, ...)
}
\arguments{
\item{object}{A \code{"GeDSboost"} class object.}

\item{...}{Further arguments (ignored).}
}
\value{
An integer indicating the total number of boosting iterations.
}
\description{
Method for \code{N.boost.iter} that returns the number of boosting iterations 
used in fitting a \code{"GeDSboost"} class object.
}
