% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GeDSMethod.R
\name{plot,GeDSboost-method}
\alias{plot,GeDSboost-method}
\alias{plot.GeDSboost}
\title{Plot method for GeDSboost objects.}
\usage{
\S4method{plot}{GeDSboost}(x, n = 3L, ...)
}
\arguments{
\item{x}{a \code{\link{GeDSboost-class}} object from which the GeDSboost fit(s) should
be extracted.}

\item{n}{integer value (2, 3 or 4) specifying the order (\eqn{=} degree
\eqn{+ 1}) of the FGB-GeDS fit.}

\item{...}{further arguments to be passed to the
\code{\link[graphics]{plot.default}} function.}
}
\description{
Plot method for GeDSboost object with a single base-learner (i.e. non-additive model).
}
