% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods_GeDSboost-GeDSgam.R
\name{predict.GeDSboost}
\alias{predict.GeDSboost}
\alias{predict.GeDSgam}
\title{Predict method for GeDSboost and GeDSgam objects}
\usage{
\method{predict}{GeDSboost}(object, newdata, n = 2L, ...)

\method{predict}{GeDSgam}(object, newdata, n = 2L, ...)
}
\arguments{
\item{object}{The \code{\link{GeDSboost-class}} or \code{\link{GeDSgam-class}} object.}

\item{newdata}{An optional data frame for prediction.}

\item{n}{The order of the GeDS fit (2 for linear, 3 for quadratic, and 4 for cubic). 
Default is 2.}

\item{...}{potentially further arguments.}
}
\value{
Numeric vector of predictions.
}
\description{
This method computes predictions from GeDSboost and GeDSgam objects. 
It is designed to be user-friendly and accommodate different orders of the GeDS fit.
}
