% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{shapeConstrain}
\alias{shapeConstrain}
\title{Apply Shape Constraints to GeDS Fits}
\usage{
shapeConstrain(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Further arguments passed to methods.}
}
\value{
An object of the same general class as \code{object}, with the selected
fit updated according to the requested shape constraint.
}
\description{
Generic function for imposing shape constraints on fitted model objects.
}
\details{
This generic function is currently implemented for selected \code{"GeDS"},
\code{"GeDSboost"} and \code{"GeDSgam"} class objects.
}
\seealso{
\code{\link{NGeDS}}, \code{\link{coef.GeDS}},
\code{\link{knots.GeDS}}, \code{\link{predict.GeDS}}
}
