% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods_GeDSboost-GeDSgam.R
\name{varimp_GeDS}
\alias{varimp_GeDS}
\title{Variable Importance for GeDS Boosting Model}
\usage{
varimp_GeDS(object, initial_model = TRUE)
}
\arguments{
\item{object}{An object of class \code{\link{GeDSboost-Class}}, which contains the results of the GeDS boosting model.}

\item{initial_model}{Logical, indicates whether to include the initial model in the importance calculation in the case of GeDS initial learner. 
If \code{TRUE}, it includes the initial model's risk difference; otherwise, it is set to 0. Default is \code{TRUE}.}
}
\value{
A numeric vector of variable importances for each base-learner, with class "varimp". 
        Attributes include "selfreqs", which gives the proportion of models where the i-th learner was selected, 
        and "variable_names", which provides the ordered names of the variables.
}
\description{
This function calculates the importance of each base-learner in the final prediction of the 
Functional Gradient Boosting model that uses Geometrically Designed splines (GeDS) as base-learners.
}
\examples{
\dontrun{
# Load packages
library(GeDS)
library(TH.data)
set.seed(290875)
data("bodyfat", package = "TH.data")
data = bodyfat
Gmodboost <- NGeDSboost(formula = DEXfat ~ f(hipcirc) + f(kneebreadth) + f(anthro3a),
data = data, initial_learner = NULL, internal_knots = 500,
max_iterations = 100, q_boost = 2, phi_boost_exit = 0.995,
shrinkage=1, phi=0.99, normalize_data = FALSE)
mse_ngedsboost1 <- mean((data$DEXfat-Gmodboost$predictions$pred_linear)^2)
mse_ngedsboost2 <- mean((data$DEXfat-Gmodboost$predictions$pred_quadratic)^2)
mse_ngedsboost3 <- mean((data$DEXfat-Gmodboost$predictions$pred_cubic)^2)

# Print
cat("\n", "MEAN SQUARED ERROR", "\n",
"Linear NGeDSboost:", mse_ngedsboost1, "\n",
"Quadratic NGeDSboost:", mse_ngedsboost2, "\n",
"Cubic NGeDSboost:", mse_ngedsboost3, "\n")
# Variable importance
varimp <- varimp_GeDS(Gmodboost)
print(varimp)
plot(varimp)
}

}
