% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R, R/S3methods_GeDSboost-GeDSgam.R
\name{deviance.GeDS}
\alias{deviance.GeDS}
\alias{deviance.GeDSboost}
\alias{deviance.GeDSgam}
\title{Deviance Method for GeDS, GeDSboost, GeDSgam}
\usage{
\method{deviance}{GeDS}(object, n = 3L, ...)

\method{deviance}{GeDSboost}(object, n = 3L, ...)

\method{deviance}{GeDSgam}(object, n = 3L, ...)
}
\arguments{
\item{object}{The \code{"GeDS"} class, \code{\link{NGeDSboost}} or
\code{\link{NGeDSgam}} object from which the deviance should be extracted.}

\item{n}{Integer value (2, 3 or 4) specifying the order (\eqn{=} degree
\eqn{+ 1}) of the GeDS/GeDSboost/GeDSgam fit whose deviance should be
extracted. By default equal to \code{3L}; non-integer values will be passed
to the function \code{\link{as.integer}}.}

\item{...}{Potentially further arguments (required by the definition of the
generic function). These will be ignored, but with a warning.}
}
\value{
A numeric value corresponding to the  deviance of the selected
GeDS/GeDSboost/GeDSgam fit.
}
\description{
Method for the function \code{\link[stats]{deviance}} that allows the user to
extract  the value of the deviance corresponding to a selected GeDS, GeDSboost
or GeDSgam fit from a \code{"GeDS"} class,
\code{\link{NGeDSboost}} or \code{\link{NGeDSgam}} object.
}
\details{
This is a method for the function \code{\link[stats]{deviance}}. As
\code{"GeDS"} class, \code{\link{NGeDSboost}} and
\code{\link{NGeDSgam}} objects contain three different fits (linear,
quadratic and cubic), it is possible to specify the order of the GeDS fit
for which the deviance is required via the input argument \code{n}.
}
\seealso{
\code{\link[stats]{deviance}} for the standard definition;
\code{\link{GGeDS}} for examples.
}
