% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R, R/S3methods_GeDSboost-GeDSgam.R
\name{family.GeDS}
\alias{family.GeDS}
\alias{family.GeDSgam}
\alias{family.GeDSboost}
\title{Extract Family from a GeDS, GeDSgam, GeDSboost Object}
\usage{
\method{family}{GeDS}(object, ...)

\method{family}{GeDSgam}(object, ...)

\method{family}{GeDSboost}(object, ...)
}
\arguments{
\item{object}{A \code{"GeDS"}, \code{"GeDSgam"} or \code{"GeDSboost"} class object.}

\item{...}{Further arguments (ignored).}
}
\value{
An object of class \code{\link[stats]{family}} describing the 
distribution and link function used in the GeDS fit.
}
\description{
Method for \code{\link[stats]{family}} that returns the error distribution 
family used in the fitted GeDS model.
}
\seealso{
\code{\link[stats]{family}}
}
