% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeDSClass.R
\docType{class}
\name{GeDSgam-class}
\alias{GeDSgam-class}
\alias{GeDSgam-Class}
\title{GeDSgam Class}
\description{
A class representing a fitted GeDSgam object returned by function \code{\link{NGeDSgam}}.
}
\section{Slots}{

\describe{
\item{\code{formula}}{A formula object representing the relationship between the variables used in the model.}

\item{\code{args}}{A list containing:
\describe{
  \item{base_learners}{List of base learners used in the model.}
  \item{family}{Family of the model.}
  \item{normalize_data}{Indicator whether the data should be normalized.}
  \item{outcome}{Outcome variable of the model.}
  \item{predictors}{Predictors used in the model.}
}}

\item{\code{final_model}}{The final GeDSgam model object containing:
\describe{
  \item{base_learners}{List of base learners used in the model. Each base learner contains:
    \describe{
      \item{coefficients}{Coefficients of the base learner.}
      \item{cubic.int.knots}{Knots used in the cubic fit.}
      \item{knots}{General knots used in the model.}
      \item{quadratic.int.knots}{Knots used in the quadratic fit.}
    }
  }
  \item{Y_hat}{Predicted values.}
  \item{model_name}{Name of the model.}
  \item{Quadratic.Fit}{Results of the quadratic fit.}
  \item{Cubic.Fit}{Results of the cubic fit.}
}}

\item{\code{predictions}}{A list containing:
\describe{
  \item{pred_linear}{Predictions from the linear fit.}
  \item{pred_quadratic}{Predictions from the quadratic fit.}
  \item{pred_cubic}{Predictions from the cubic fit.}
}}
}}

\references{
Dimitrova, D.S., Kaishev, V.K., Lattuada A. and Verrall, R.J. (2017).
Geometrically designed, variable knot splines in Generalized (Non-)Linear Models.
Available at \href{http://openaccess.city.ac.uk/18460/}{openaccess.city.ac.uk}
}
\seealso{
\code{\link{NGeDSgam}} which generates objects of this class.
}
