% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods_GeDSboost-GeDSgam.R
\name{n.boost.iter}
\alias{n.boost.iter}
\alias{n.boost.iter.GeDSboost}
\title{Extract Number of Boosting Iterations from a GeDSboost Object}
\usage{
\method{n.boost.iter}{GeDSboost}(object, ...)
}
\arguments{
\item{object}{a \code{\link{GeDSboost-class}} object.}

\item{...}{further arguments (ignored).}
}
\value{
An integer indicating the number of boosting iterations used.
}
\description{
Method for \code{n.boost.iter} that returns the number of boosting iterations 
used in fitting a \code{\link{GeDSboost-class}} object.
}
