% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGeDSgam.R
\name{NGeDSgam}
\alias{NGeDSgam}
\title{NGeDSgam: Local Scoring Algorithm with GeD Splines in Backfitting}
\usage{
NGeDSgam(
  formula,
  data,
  weights = NULL,
  offset = NULL,
  normalize_data = FALSE,
  family = "gaussian",
  min_iterations = 0,
  max_iterations = 100,
  phi_gam_exit = 0.995,
  q_gam = 2,
  beta = 0.5,
  phi = 0.99,
  internal_knots = 500,
  q = 2,
  higher_order = TRUE
)
}
\arguments{
\item{formula}{a formula specifying the model structure. It is expected to be in the
format `Y ~ f(x1) + f(x2) + ...`.}

\item{data}{a data frame containing the variables referenced in the formula.}

\item{weights}{optional vector of weights for each observation.}

\item{offset}{optional offset to be added to the linear predictor during the backfitting.}

\item{normalize_data}{a logical that defines whether the data is normalized before fitting the model or not.}

\item{family}{a character string indicating the response distribution and link function to be used. Default is "gaussian".
This should be a character or a family object. Accepted characters are "gaussian" and "binomial".}

\item{min_iterations}{Minimum number of iterations in the algorithm. Default is 0.}

\item{max_iterations}{Maximum number of iterations in the algorithm. Default is 100.}

\item{phi_gam_exit}{Convergence threshold for local-scoring. The algorithm stops when the relative change in the deviance is below this threshold.
Default is 1e-3.}

\item{q_gam}{numeric parameter which allows to fine-tune the stopping rule of stage A of GeDS, by default equal to 2.}

\item{beta}{numeric parameter in the interval \eqn{[0,1]}
tuning the knot placement in stage A of GeDS. See details.}

\item{phi}{numeric parameter in the interval \eqn{[0,1]} specifying the threshold for
the stopping rule  (model selector) in stage A of GeDS. See also \code{stoptype} and details below.}

\item{internal_knots}{maximum number of internal knots to be added by the GeDS learners within the backfitting iterations. Default is 500.}

\item{q}{numeric parameter which allows to fine-tune the stopping rule of stage A of GeDS, by default equal to 2.}

\item{higher_order}{a logical that defines whether to compute the higher order fits (quadratic and cubic).}
}
\value{
A list of class "GeDSgam" containing:
  * formula: The formula provided as input.
  * args: List of arguments and relevant data used internally.
  * final_model: The final fitted model after local scoring.
  * predictions: Predictions for linear, quadratic, and cubic fits.
}
\description{
Implements the Local Scoring Algorithm (Hastie and Tibshirani (1986)),
applying GeD splines to fit the targets within the backfitting iterations.
}
\examples{
# Load package
library(GeDS) 

data(airquality) 
data = na.omit(airquality)
data$Ozone <- data$Ozone^(1/3)

formula = Ozone ~ f(Solar.R) + f(Wind, Temp)
Gmod_gam <- NGeDSgam(formula = formula, data = data,
phi_gam_exit = 0.995, phi = 0.995, q = 2)
mse_ngedsgam1 <- mean((data$Ozone - Gmod_gam$predictions$pred_linear)^2)
mse_ngedsgam2 <- mean((data$Ozone - Gmod_gam$predictions$pred_quadratic)^2)
mse_ngedsgam3 <- mean((data$Ozone - Gmod_gam$predictions$pred_cubic)^2)

cat("\n", "MEAN SQUARED ERROR", "\n",
"Linear NGeDSgam:", mse_ngedsgam1, "\n",
"Quadratic NGeDSgam:", mse_ngedsgam2, "\n",
"Cubic NGeDSgam:", mse_ngedsgam3, "\n")

}
\references{
Hastie, T., & Tibshirani, R. (1986). Generalized Additive Models. Statistical Science, 1(3), 297–310. http://www.jstor.org/stable/2245459
}
\seealso{
\link{gam}, \link{glm}
}
