% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GeDSMethod.R
\name{plot,GeDSgam-method}
\alias{plot,GeDSgam-method}
\alias{plot.GeDSgam}
\alias{plot,GeDSgam,ANY-method}
\title{Plot method for GeDSgam objects.}
\usage{
\S4method{plot}{GeDSgam,ANY}(x, n = 3L, ...)
}
\arguments{
\item{x}{a \code{\link{GeDSgam-class}} object from which the GeDSgam fit(s) should
be extracted.}

\item{n}{integer value (2, 3 or 4) specifying the order (\eqn{=} degree
\eqn{+ 1}) of the GAM-GeDS fit.}

\item{...}{further arguments to be passed to the
\code{\link[graphics]{plot.default}} function.}
}
\description{
Plots on the linear predictor scale the component functions of a GeDSgam
object fitted using \code{\link{NGeDSgam}}.
}
