% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods_GeDSboost-GeDSgam.R
\name{plot.GeDSboost}
\alias{plot.GeDSboost}
\title{Plot method for non-additive GeDSboost}
\usage{
\method{plot}{GeDSboost}(object, n = 3L, ...)
}
\arguments{
\item{object}{the  \code{\link{GeDSboost-class}}}

\item{n}{integer value (2, 3 or 4) specifying the order (\eqn{=} degree
\eqn{+ 1}) of the FGB-GeDS fit whose coefficients should be
extracted.}

\item{...}{potentially further arguments (required by the definition of the
generic function). They will be ignored, but with a warning.}
}
\description{
Plot method for GeDSboost objects. Plots GeDSboost fits.
}
