% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R, R/S3methods_GeDSboost-GeDSgam.R
\name{summary.GeDS}
\alias{summary.GeDS}
\alias{summary.GeDSboost}
\alias{summary.GeDSgam}
\title{Summary Method for GeDS, GeDSgam, GeDSboost}
\usage{
\method{summary}{GeDS}(object, ...)

\method{summary}{GeDSgam}(object, ...)

\method{summary}{GeDSboost}(object, ...)
}
\arguments{
\item{object}{The \code{"GeDS"}, \code{"GeDSgam"} or \code{"GeDSboost"} object
for which the main information should be printed on screen.}

\item{...}{Potentially further arguments (required by the definition of the
generic function).}
}
\description{
Method for the generic function \code{\link[base]{summary}} that allows you to
print on screen the main information related to a fitted \code{"GeDS"},
\code{"GeDSgam"} or \code{"GeDSboost"} model.
Similar to \code{\link{print.GeDS}} but with some extra detail.
}
\seealso{
\code{\link{print.GeDS}}
}
