% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods_GeDSboost-GeDSgam.R
\name{visualize_boosting}
\alias{visualize_boosting}
\title{Visualize Boosting Process}
\usage{
visualize_boosting(M, Gmodboost)
}
\arguments{
\item{M}{Numeric, specifies the iteration number.}

\item{Gmodboost}{An object of class  \code{\link{GeDSboost-Class}}, containing the results of the GeDS boosting model.}
}
\description{
This function presents  the fit over the data at the start of each boosting iteration and the subsequent fit on the residuals.
}
\examples{
\dontrun{
# Load packages
library(GeDS)

# Generate a data sample for the response variable
# Y and the single covariate X
set.seed(123)
N <- 500
f_1 <- function(x) (10*x/(1+100*x^2))*4+4
X <- sort(runif(N, min = -2, max = 2))
# Specify a model for the mean of Y to include only a component
# non-linear in X, defined by the function f_1
means <- f_1(X)
# Add (Normal) noise to the mean of Y
Y <- rnorm(N, means, sd = 0.2)
data = data.frame(X, Y)
Gmodboost <- NGeDSboost(Y ~ f(X), data = data, initial_learner = "GeDS", int.knots_init = 2,
internal_knots = 500, q_boost = 2, phi_boost_exit = 0.995, shrinkage = 1,
phi = 0.99, normalize_data = TRUE, family = mboost::Gaussian())

# Plot
plot(X, Y, pch=20, col=c("darkgrey"))
lines(X, sapply(X, f_1), col = "black", lwd = 2)
lines(X, Gmodboost$predictions$pred_linear, col = "green4", lwd = 2)
lines(X, Gmodboost$predictions$pred_quadratic, col="red", lwd=2)
lines(X, Gmodboost$predictions$pred_cubic, col="purple", lwd=2)
legend("topright",
legend = c("Order 2 (degree=1)", "Order 3 (degree=2)", "Order 4 (degree=3)"),
col = c("green4", "red", "purple"),
lty = c(1, 1),
lwd = c(2, 2, 2),
cex = 0.75,
bty="n",
bg = "white")
# Visualize boosting iterations
par(mfrow=c(2,2))
visualize_boosting(0, Gmodboost)
visualize_boosting(1, Gmodboost)
par(mfrow=c(1,1))
}

}
